SPARK-1555 enable ec2/spark_ec2.py to stop/delete cluster non-interactively
Added new command-line option: --assume-yes

It is a boolean flag that defaults to false.

Before asking for user input on stop/delete cluster, we check ops.assume_yes.
If opts.assume_yes is True, we do not ask for user input and instead assume
the answer is yes.

If opts.assume_yes is False, we ask for user input and take the answer from
there.